fix/add error if missing where

+++ DORM/Database/SQLClasses/Delete.php
@@ -27,6 +27,10 @@ class Delete
 
     public function __toString()
     {
+        if ($this->where === null) {
+            throw new \Exception("SECURITY: UPDATE needs a WHERE argument");
+        }
+        
         switch ($this->sqlType) {
             case 'mysql':
                 return "DELETE FROM " . $this->table
